fix(computer-use): normalize keyboard key and modifier aliases on windows
