+++ libs/computer-use/pkg/computeruse/keyboard_normalization.go
@@ -1,5 +1,3 @@
-//go:build linux
-
 // Copyright Daytona Platforms Inc.
 // SPDX-License-Identifier: AGPL-3.0
